refactor(chat-composer): address review — share composer provider actions
- extract ProviderActionHandlers + emptyActions into variants/shared/
  composerProviderActions (both variants drifted: addNewTopic was
  (payload?) => void in chat vs () => void in agent — now one signature)
- relocate AddNewTopicPayload there too and re-export it from pages/home/types,
  so the composer no longer imports upward into pages/ (fixes the
  components -> pages layer inversion)
