+++ src/renderer/pages/home/types.ts
@@ -1,3 +1,3 @@
-export interface AddNewTopicPayload {
-  assistantId?: string | null
-}
+// Defined in the composer layer; re-exported here for page-side consumers so the
+// composer doesn't import upward into `pages/`.
+export type { AddNewTopicPayload } from '@renderer/components/chat/composer/variants/shared/composerProviderActions'
